refactor(runtime): extract layout renderer helpers
